final_project/requirements.txt
flask==2.3.2
ibm-watson==7.0.0
python-dotenv
pyJWT==2.7.0
pylint